Move show() method to _BaseViewer
